pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "style-search"
version = "0.1.0"
description = "Scrape doomp images and explore style-space with embeddings"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14",
    "chromadb>=1.4",
    "click>=8.3",
    "fastapi>=0.128",
    "llvmlite>=0.46",
    "numba>=0.63",
    "numpy>=1.22,<2.4",
    "open-clip-torch>=3.2",
    "pillow>=12.1",
    "requests>=2.32",
    "rich>=14.3",
    "torch>=2.10",
    "torchvision>=0.25",
    "umap-learn>=0.5",
    "uvicorn>=0.40",
]

[project.scripts]
scrape-doomp = "style_search.scrape_doomp:main"
embed-images = "style_search.embed_images:main"
style-api = "style_search.api:main"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
